requirements.txt
aiogram
aiogram-dialog
requests
deepgram-sdk
